default rendering

+++ shaders/marching_cube/draw.frag
@@ -3,5 +3,5 @@
 layout(location=0) in vec3 inColor;
 layout(location=0) out vec4 outColor;
 void main(){
-	outColor = uvec4(inColor, 1.0);
+	outColor = vec4(inColor, 1.0);
 }
